+++ examples/basic-2/generated-client/accounts/Counter.ts
@@ -1,7 +1,7 @@
 import { PublicKey, Connection } from "@solana/web3.js"
 import BN from "bn.js" // eslint-disable-line @typescript-eslint/no-unused-vars
 import * as borsh from "@project-serum/borsh" // eslint-disable-line @typescript-eslint/no-unused-vars
-import { PROGRAM_ID } from "../programId"
+import { PROGRAM_ID, programIdOverride } from "../programId"
 
 export interface CounterFields {
   authority: PublicKey
@@ -35,12 +35,13 @@ export class Counter {
     c: Connection,
     address: PublicKey
   ): Promise<Counter | null> {
+    const programId = (programIdOverride && programIdOverride()) || PROGRAM_ID
     const info = await c.getAccountInfo(address)
 
     if (info === null) {
       return null
     }
-    if (!info.owner.equals(PROGRAM_ID)) {
+    if (!info.owner.equals(programId)) {
       throw new Error("account doesn't belong to this program")
     }
 
@@ -51,13 +52,14 @@ export class Counter {
     c: Connection,
     addresses: PublicKey[]
   ): Promise<Array<Counter | null>> {
+    const programId = (programIdOverride && programIdOverride()) || PROGRAM_ID
     const infos = await c.getMultipleAccountsInfo(addresses)
 
     return infos.map((info) => {
       if (info === null) {
         return null
       }
-      if (!info.owner.equals(PROGRAM_ID)) {
+      if (!info.owner.equals(programId)) {
         throw new Error("account doesn't belong to this program")
       }