+++ examples/tic-tac-toe/generated-client/accounts/Game.ts
@@ -2,7 +2,7 @@ import { PublicKey, Connection } from "@solana/web3.js"
 import BN from "bn.js" // eslint-disable-line @typescript-eslint/no-unused-vars
 import * as borsh from "@project-serum/borsh" // eslint-disable-line @typescript-eslint/no-unused-vars
 import * as types from "../types" // eslint-disable-line @typescript-eslint/no-unused-vars
-import { PROGRAM_ID } from "../programId"
+import { PROGRAM_ID, programIdOverride } from "../programId"
 
 export interface GameFields {
   players: Array<PublicKey>
@@ -43,12 +43,13 @@ export class Game {
   }
 
   static async fetch(c: Connection, address: PublicKey): Promise<Game | null> {
+    const programId = (programIdOverride && programIdOverride()) || PROGRAM_ID
     const info = await c.getAccountInfo(address)
 
     if (info === null) {
       return null
     }
-    if (!info.owner.equals(PROGRAM_ID)) {
+    if (!info.owner.equals(programId)) {
       throw new Error("account doesn't belong to this program")
     }
 
@@ -59,13 +60,14 @@ export class Game {
     c: Connection,
     addresses: PublicKey[]
   ): Promise<Array<Game | null>> {
+    const programId = (programIdOverride && programIdOverride()) || PROGRAM_ID
     const infos = await c.getMultipleAccountsInfo(addresses)
 
     return infos.map((info) => {
       if (info === null) {
         return null
       }
-      if (!info.owner.equals(PROGRAM_ID)) {
+      if (!info.owner.equals(programId)) {
         throw new Error("account doesn't belong to this program")
       }