+++ tests/example-program-gen/exp/accounts/State2.ts
@@ -2,7 +2,7 @@ import { PublicKey, Connection } from "@solana/web3.js"
 import BN from "bn.js" // eslint-disable-line @typescript-eslint/no-unused-vars
 import * as borsh from "@project-serum/borsh" // eslint-disable-line @typescript-eslint/no-unused-vars
 import * as types from "../types" // eslint-disable-line @typescript-eslint/no-unused-vars
-import { PROGRAM_ID } from "../programId"
+import { PROGRAM_ID, programIdOverride } from "../programId"
 
 export interface State2Fields {
   vecOfOption: Array<BN | null>
@@ -31,12 +31,13 @@ export class State2 {
     c: Connection,
     address: PublicKey
   ): Promise<State2 | null> {
+    const programId = (programIdOverride && programIdOverride()) || PROGRAM_ID
     const info = await c.getAccountInfo(address)
 
     if (info === null) {
       return null
     }
-    if (!info.owner.equals(PROGRAM_ID)) {
+    if (!info.owner.equals(programId)) {
       throw new Error("account doesn't belong to this program")
     }
 
@@ -47,13 +48,14 @@ export class State2 {
     c: Connection,
     addresses: PublicKey[]
   ): Promise<Array<State2 | null>> {
+    const programId = (programIdOverride && programIdOverride()) || PROGRAM_ID
     const infos = await c.getMultipleAccountsInfo(addresses)
 
     return infos.map((info) => {
       if (info === null) {
         return null
       }
-      if (!info.owner.equals(PROGRAM_ID)) {
+      if (!info.owner.equals(programId)) {
         throw new Error("account doesn't belong to this program")
       }