+++ tests/example-program-gen/exp/instructions/initialize.ts
@@ -2,7 +2,7 @@ import { TransactionInstruction, PublicKey, AccountMeta } from "@solana/web3.js"
 import BN from "bn.js" // eslint-disable-line @typescript-eslint/no-unused-vars
 import * as borsh from "@project-serum/borsh" // eslint-disable-line @typescript-eslint/no-unused-vars
 import * as types from "../types" // eslint-disable-line @typescript-eslint/no-unused-vars
-import { PROGRAM_ID } from "../programId"
+import { PROGRAM_ID, programIdOverride } from "../programId"
 
 export interface InitializeAccounts {
   /** State account */
@@ -17,6 +17,7 @@ export interface InitializeAccounts {
 }
 
 export function initialize(accounts: InitializeAccounts) {
+  const programId = (programIdOverride && programIdOverride()) || PROGRAM_ID
   const keys: Array<AccountMeta> = [
     { pubkey: accounts.state, isSigner: true, isWritable: true },
     { pubkey: accounts.nested.clock, isSigner: false, isWritable: false },
@@ -26,6 +27,6 @@ export function initialize(accounts: InitializeAccounts) {
   ]
   const identifier = Buffer.from([175, 175, 109, 31, 13, 152, 155, 237])
   const data = identifier
-  const ix = new TransactionInstruction({ keys, programId: PROGRAM_ID, data })
+  const ix = new TransactionInstruction({ keys, programId, data })
   return ix
 }