+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Features
+
+- Added programIdOverride to programId.ts so that program IDs can be overriden dynamically by the app ([#55](https://github.com/kklas/anchor-client-gen/pull/55))
+
 ## [v0.25.0]
 
 ### Features